Cleans up logic in order details
Removed some potential issue areas associated with setting the shipping
address for the order details. Previously this was being pulled from
the shipment, now it is pulled from the Order object which should result
in more consistent data. Also added in an isPresent() check when accessing
the repository for the shipment data.